moving setup into conftest.py
