+++ tests/feature_test.py
@@ -3,15 +3,9 @@ import sys
 
 import mock
 import pytest
+import traceback
+
 from click.testing import CliRunner
-from swsssdk import ConfigDBConnector
-
-test_path = os.path.dirname(os.path.abspath(__file__))
-modules_path = os.path.dirname(test_path)
-sys.path.insert(0, modules_path)
-
-import show.main as show
-import config.main as config
 
 show_feature_status_output="""\
 Feature     State     AutoRestart
@@ -76,32 +70,25 @@ Feature    AutoRestart
 bgp        disabled
 """
 
-@pytest.fixture
-def setup_config_db():
-    config.asic_type = mock.MagicMock(return_value = "broadcom")
-    config._get_device_type = mock.MagicMock(return_value = "ToRRouter")
-
-    config_db = ConfigDBConnector()
-    config_db.connect()
-
-    show.config_db = config_db
-    config.config_db = config_db
-
-
 class TestFeature(object):
     @classmethod
     def setup_class(cls):
         print("SETUP")
+        os.environ["UTILITIES_UNIT_TESTING"] = "1"
 
     def test_show_feature_status(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(show.cli.commands["feature"].commands["status"], [])
         print(result.exit_code)
         print(result.output)
+        print(result.exception)
+        traceback.print_tb(result.exc_info[2])
         assert result.exit_code == 0
         assert result.output == show_feature_status_output
 
     def test_show_bgp_feature_status(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(show.cli.commands["feature"].commands["status"], ["bgp"])
         print(result.exit_code)
@@ -110,6 +97,7 @@ class TestFeature(object):
         assert result.output == show_feature_bgp_status_output
 
     def test_show_unknown_feature_status(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(show.cli.commands["feature"].commands["status"], ["foo"])
         print(result.exit_code)
@@ -117,6 +105,7 @@ class TestFeature(object):
         assert result.exit_code == 1
 
     def test_show_feature_autorestart(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(show.cli.commands["feature"].commands["autorestart"], [])
         print(result.exit_code)
@@ -125,6 +114,7 @@ class TestFeature(object):
         assert result.output == show_feature_autorestart_output
 
     def test_show_bgp_autorestart_status(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(show.cli.commands["feature"].commands["autorestart"], ["bgp"])
         print(result.exit_code)
@@ -133,6 +123,7 @@ class TestFeature(object):
         assert result.output == show_feature_bgp_autorestart_output
 
     def test_show_unknown_autorestart_status(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(show.cli.commands["feature"].commands["autorestart"], ["foo"])
         print(result.exit_code)
@@ -140,6 +131,7 @@ class TestFeature(object):
         assert result.exit_code == 1
 
     def test_config_bgp_feature_state(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(config.config.commands["feature"].commands["state"], ["bgp", "disabled"])
         print(result.exit_code)
@@ -150,6 +142,7 @@ class TestFeature(object):
         assert result.output == show_feature_bgp_disabled_status_output
 
     def test_config_bgp_autorestart(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(config.config.commands["feature"].commands["autorestart"], ["bgp", "disabled"])
         print(result.exit_code)
@@ -160,6 +153,7 @@ class TestFeature(object):
         assert result.output == show_feature_bgp_disabled_autorestart_output
 
     def test_config_unknown_feature(self, setup_config_db):
+        (config, show) = setup_config_db
         runner = CliRunner()
         result = runner.invoke(config.config.commands["feature"].commands['state'], ["foo", "enabled"])
         print(result.output)
